move logic to EditSqlModal
