+++ packages/front-end/components/SchemaBrowser/EditSqlModal.tsx
@@ -28,7 +28,7 @@ export interface Props {
   close: () => void;
   requiredColumns: Set<string>;
   placeholder?: string;
-  validateResponse?: (response: TestQueryRow) => void;
+  validateResponseOverride?: (response: TestQueryRow) => void;
   templateVariables?: {
     eventName?: string;
     valueColumn?: string;
@@ -42,7 +42,7 @@ export default function EditSqlModal({
   requiredColumns,
   placeholder = "",
   datasourceId,
-  validateResponse,
+  validateResponseOverride,
   templateVariables,
 }: Props) {
   const [
@@ -62,6 +62,21 @@ export default function EditSqlModal({
   const [cursorData, setCursorData] = useState<null | CursorData>(null);
   const [testingQuery, setTestingQuery] = useState(false);
 
+  const validateRequiredColumns = (result: TestQueryRow) => {
+    if (!result) return;
+
+    const requiredColumnsArray = Array.from(requiredColumns);
+    const missingColumns = requiredColumnsArray.filter(
+      (col) => !((col as string) in result)
+    );
+
+    if (missingColumns.length > 0) {
+      throw new Error(
+        `You are missing the following columns: ${missingColumns.join(", ")}`
+      );
+    }
+  };
+
   const runTestQuery = async (sql: string) => {
     validateSQL(sql, []);
     setTestQueryResults(null);
@@ -74,8 +89,12 @@ export default function EditSqlModal({
       }),
     });
 
-    if (res.results?.length && validateResponse) {
-      validateResponse(res.results[0]);
+    if (res.results?.length) {
+      if (validateResponseOverride) {
+        validateResponseOverride(res.results[0]);
+      } else {
+        validateRequiredColumns(res.results[0]);
+      }
     }
 
     return res;
@@ -92,6 +111,7 @@ export default function EditSqlModal({
     }
     setTestingQuery(false);
   };
+
   const datasource = getDatasourceById(datasourceId);
   const supportsSchemaBrowser =
     datasource?.properties?.supportsInformationSchema;
